Use modern way of loading the bundle's config by extending AbstractBundle instead of Bundle
