+++ src/DependencyInjection/EasyAdminExtension.php
@@ -2,82 +2,19 @@
 
 namespace EasyCorp\Bundle\EasyAdminBundle\DependencyInjection;
 
-use EasyCorp\Bundle\EasyAdminBundle\Attribute\AdminRoute;
-use EasyCorp\Bundle\EasyAdminBundle\Contracts\Action\ActionsExtensionInterface;
-use EasyCorp\Bundle\EasyAdminBundle\Contracts\Controller\CrudControllerInterface;
-use EasyCorp\Bundle\EasyAdminBundle\Contracts\Controller\DashboardControllerInterface;
-use EasyCorp\Bundle\EasyAdminBundle\Contracts\Field\FieldConfiguratorInterface;
-use EasyCorp\Bundle\EasyAdminBundle\Contracts\Filter\FilterConfiguratorInterface;
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Extension\Extension;
-use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
-use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
+use EasyCorp\Bundle\EasyAdminBundle\EasyAdminBundle;
 
 /**
  * @author Javier Eguiluz <javier.eguiluz@gmail.com>
+ *
+ * @deprecated since 5.5.1, use \EasyCorp\Bundle\EasyAdminBundle\EasyAdminBundle instead
  */
-class EasyAdminExtension extends Extension implements PrependExtensionInterface
+class EasyAdminExtension
 {
-    public const TAG_CRUD_CONTROLLER = 'ea.crud_controller';
-    public const TAG_DASHBOARD_CONTROLLER = 'ea.dashboard_controller';
-    public const TAG_ADMIN_ROUTE_CONTROLLER = 'ea.admin_route_controller';
-    public const TAG_FIELD_CONFIGURATOR = 'ea.field_configurator';
-    public const TAG_FILTER_CONFIGURATOR = 'ea.filter_configurator';
-    public const TAG_ACTIONS_EXTENSION = 'ea.actions_extension';
-
-    public function load(array $configs, ContainerBuilder $container): void
-    {
-        $container->registerAttributeForAutoconfiguration(AdminRoute::class,
-            // @phpstan-ignore-next-line argument.type The reflection subtypes specify where the attribute can be used
-            static function (Definition $definition, AdminRoute $attribute, \ReflectionClass|\ReflectionMethod $reflection): void {
-                $definition->addTag(self::TAG_ADMIN_ROUTE_CONTROLLER);
-            });
-
-        $container->registerForAutoconfiguration(DashboardControllerInterface::class)
-            ->addTag(self::TAG_DASHBOARD_CONTROLLER);
-
-        $container->registerForAutoconfiguration(CrudControllerInterface::class)
-            ->addTag(self::TAG_CRUD_CONTROLLER);
-
-        $container->registerForAutoconfiguration(FieldConfiguratorInterface::class)
-            ->addTag(self::TAG_FIELD_CONFIGURATOR);
-
-        $container->registerForAutoconfiguration(FilterConfiguratorInterface::class)
-            ->addTag(self::TAG_FILTER_CONFIGURATOR);
-
-        $container->registerForAutoconfiguration(ActionsExtensionInterface::class)
-            ->addTag(self::TAG_ACTIONS_EXTENSION);
-
-        $loader = new PhpFileLoader($container, new FileLocator(__DIR__.'/../../config'));
-        $loader->load('services.php');
-    }
-
-    public function prepend(ContainerBuilder $builder): void
-    {
-        $builder->prependExtensionConfig('twig_component', [
-            'defaults' => [
-                'EasyCorp\\Bundle\\EasyAdminBundle\\Twig\\Component\\' => [
-                    'template_directory' => '@EasyAdmin/components/',
-                    'name_prefix' => 'ea',
-                ],
-            ],
-        ]);
-
-        /** @var string $projectDir */
-        $projectDir = $builder->getParameter('kernel.project_dir');
-
-        $bundleTemplatesOverrideDir = $projectDir.'/templates/bundles/EasyAdminBundle/';
-        $builder->prependExtensionConfig('twig', [
-            'paths' => is_dir($bundleTemplatesOverrideDir)
-                ? [
-                    'templates/bundles/EasyAdminBundle/' => 'ea',
-                    \dirname(__DIR__).'/../templates/' => 'ea',
-                ]
-                : [
-                    \dirname(__DIR__).'/../templates/' => 'ea',
-                ],
-        ]);
-    }
+    public const TAG_CRUD_CONTROLLER = EasyAdminBundle::TAG_CRUD_CONTROLLER;
+    public const TAG_DASHBOARD_CONTROLLER = EasyAdminBundle::TAG_DASHBOARD_CONTROLLER;
+    public const TAG_ADMIN_ROUTE_CONTROLLER = EasyAdminBundle::TAG_ADMIN_ROUTE_CONTROLLER;
+    public const TAG_FIELD_CONFIGURATOR = EasyAdminBundle::TAG_FIELD_CONFIGURATOR;
+    public const TAG_FILTER_CONFIGURATOR = EasyAdminBundle::TAG_FILTER_CONFIGURATOR;
+    public const TAG_ACTIONS_EXTENSION = EasyAdminBundle::TAG_ACTIONS_EXTENSION;
 }
